Move useResetState into hooks
